// src/order/dtos/create-order.dto.ts
import { IsNotEmpty, IsString } from "class-validator";

export class CreateOrderDto {
    first_name: string;
    last_name: string;
    email: string;
    address: string;
    country: string;
    city: string;
    zip: string;
    code: string;
    products: {
        product_id: number;
        quantity: number;
    }[];
}